norme finie

+++ minishell_m/exec/here_doc.c
@@ -6,66 +6,11 @@
 /*   By: bbouarab <bbouarab@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2013/11/18 13:37:42 by kube              #+#    #+#             */
-/*   Updated: 2026/01/24 19:28:16 by bbouarab         ###   ########.fr       */
+/*   Updated: 2026/01/24 22:20:57 by bbouarab         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
-
-char	*ignore_quotes(char *str)
-{
-	int		i;
-	int		j;
-	char	*copy;
-
-	i = 0;
-	j = 0;
-	while (str[i])
-	{
-		if (str[i] == '\'' || str[i] == '\"')
-			j++;
-		i++;
-	}
-	copy = malloc((i - j) + 1);
-	i = 0;
-	j = 0;
-	while (str[i])
-	{
-		while (str[i] == '\'' || str[i] == '\"')
-			i++;
-		if (str[i])
-			copy[j++] = str[i++];
-	}
-	return (copy[j] = 0, copy);
-}
-
-int	should_expand_here_doc(char *file)
-{
-	int	i;
-
-	i = 0;
-	while (file[i])
-	{
-		if (file[i] == '\'' || file[i] == '\"')
-			return (0);
-		i++;
-	}
-	return (1);
-}
-
-int	get_here_doc_nbr(t_token *token)
-{
-	int	i;
-
-	i = 0;
-	while (token && token->type != PIPE)
-	{
-		if (token->type == HERE_DOC)
-			i++;
-		token = token->next;
-	}
-	return (i);
-}
 
 void	get_here_doc_expand(t_token *token, t_cmd *cmd)
 {
@@ -91,17 +36,11 @@ void	get_here_doc_expand(t_token *token, t_cmd *cmd)
 	}
 }
 
-void	here_doc_expand(t_cmd *cmd, char *lim, int i)
+void	here_doc_expand_2(t_cmd *cmd, char *lim, int i)
 {
-	char	*file;
-	char	*nbr;
 	char	*expanded_line;
 
 	expanded_line = NULL;
-	nbr = ft_itoa(cmd->i);
-	file = ft_strjoin("/var/tmp/temp", nbr);
-	cmd->minishell->in_here_doc = 1;
-	cmd->redir->fd[i] = open(file, O_RDWR | O_TRUNC | O_CREAT, 0644);
 	while (cmd->minishell->in_here_doc && ft_findstr(lim, cmd->minishell->gnl))
 	{
 		if (cmd->minishell->gnl)
@@ -118,12 +57,28 @@ void	here_doc_expand(t_cmd *cmd, char *lim, int i)
 		if (!cmd->minishell->gnl || !expanded_line)
 			break ;
 	}
+}
+
+void	here_doc_expand(t_cmd *cmd, char *lim, int i)
+{
+	char	*file;
+	char	*nbr;
+
+	nbr = ft_itoa(cmd->i);
+	file = ft_strjoin("/var/tmp/temp", nbr);
+	cmd->minishell->in_here_doc = 1;
+	cmd->redir->fd[i] = open(file, O_RDWR | O_TRUNC | O_CREAT, 0644);
+	if (cmd->redir->fd[i] < 0)
+		return (perror("open"), free_ms(cmd->token, NULL, 1));
+	here_doc_expand_2(cmd, lim, i);
 	cmd->minishell->in_here_doc = 0;
 	if (cmd->minishell->gnl)
 		free(cmd->minishell->gnl);
 	cmd->minishell->gnl = get_next_line(0, 2);
 	close(cmd->redir->fd[i]);
 	cmd->redir->fd[i] = open(file, O_RDWR, 0644);
+	if (cmd->redir->fd[i] < 0)
+		return (perror("open"), free_ms(cmd->token, NULL, 1));
 	unlink(file);
 	free(file);
 	free(nbr);
